[fix] Code Refactoring by Router
