+++ .gitignore
@@ -2,3 +2,7 @@ myvenv/
 config/
 __pycache__/
 .idea/
+alembic.ini
+data/
+new_chroma_db/
+migrations/
